Move winston NodeJS transports to cli package
