+++ packages/cli/test/unit/util/loggerTransport.test.ts
@@ -3,8 +3,8 @@ import path from "node:path";
 import {Writable} from "node:stream";
 import rimraf from "rimraf";
 import {expect} from "chai";
-import {LogData, LodestarError, LogFormat, logFormats, LogLevel, WinstonLogger} from "../../../src/index.js";
-import {TransportType} from "../../../src/logger/transport.js";
+import {LodestarError, LogData, LogFormat, logFormats, LogLevel, WinstonLogger} from "@lodestar/utils";
+import {fromTransportOpts, TransportType} from "../../../src/util/loggerTransports.js";
 
 /**
  * To capture Winston output in memory
@@ -74,7 +74,9 @@ describe("winston logger", () => {
       for (const format of logFormats) {
         it(`${id} ${format} output`, async () => {
           const stream = new WritableMemory();
-          const logger = new WinstonLogger({format, hideTimestamp: true}, [{type: TransportType.stream, stream}]);
+          const logger = new WinstonLogger({format, hideTimestamp: true}, [
+            fromTransportOpts({type: TransportType.stream, stream}),
+          ]);
           logger.warn(message, context, error);
 
           expect(stream.getAsString().trim()).to.equal(output[format]);
@@ -86,7 +88,9 @@ describe("winston logger", () => {
   describe("child logger", () => {
     it("Should parse child module", async () => {
       const stream = new WritableMemory();
-      const logger = new WinstonLogger({hideTimestamp: true, module: "A"}, [{type: TransportType.stream, stream}]);
+      const logger = new WinstonLogger({hideTimestamp: true, module: "A"}, [
+        fromTransportOpts({type: TransportType.stream, stream}),
+      ]);
       const childB = logger.child({module: "B"});
       const childC = childB.child({module: "C"});
       childC.warn("test");
@@ -97,7 +101,7 @@ describe("winston logger", () => {
     it("Should log to child at a lower logLevel", () => {
       const stream = new WritableMemory();
       const logger = new WinstonLogger({hideTimestamp: true, module: "A"}, [
-        {type: TransportType.stream, stream, level: LogLevel.info},
+        fromTransportOpts({type: TransportType.stream, stream, level: LogLevel.info}),
       ]);
 
       const childB = logger.child({module: "B", level: LogLevel.debug});
@@ -119,7 +123,9 @@ describe("winston logger", () => {
     it("Should log to file", async () => {
       const filename = path.join(tmpDir, "child-logger-test.txt");
 
-      const logger = new WinstonLogger({hideTimestamp: true, module: "A"}, [{type: TransportType.file, filename}]);
+      const logger = new WinstonLogger({hideTimestamp: true, module: "A"}, [
+        fromTransportOpts({type: TransportType.file, filename}),
+      ]);
       logger.warn("test");
 
       const output = await readFileWhenExists(filename);
