+++ packages/utils/src/logger/index.ts
@@ -1,5 +1,4 @@
 export * from "./interface.js";
 export * from "./format.js";
-export * from "./transport.js";
 export * from "./winston.js";
 export {LogData} from "./json.js";
