+++ packages/utils/src/logger/winston.ts
@@ -1,15 +1,11 @@
-import {Writable} from "node:stream";
 import winston from "winston";
 import type {Logger} from "winston";
 import chalk from "chalk";
-import {defaultLogLevel, ILogger, ILoggerOptions, LogLevel, logLevelNum} from "./interface.js";
+import {defaultLogLevel, ILogger, ILoggerOptions, LogLevel} from "./interface.js";
 import {getFormat} from "./format.js";
-import {TransportOpts, TransportType, fromTransportOpts} from "./transport.js";
 import {LogData} from "./json.js";
 
 const {createLogger} = winston;
-
-const defaultTransportOpts: TransportOpts = {type: TransportType.console};
 
 type DefaultMeta = {
   module: string;
@@ -17,30 +13,18 @@ type DefaultMeta = {
 
 export class WinstonLogger implements ILogger {
   private winston: Logger;
-  private _level: LogLevel;
-  private _options: Partial<ILoggerOptions>;
-  private _transportOptsArr: TransportOpts[];
 
-  constructor(options: Partial<ILoggerOptions> = {}, transportOptsArr: TransportOpts[] = [defaultTransportOpts]) {
-    // `options.level` can override the level in the transport
-    // This is necessary for child logger opts to take effect
-    let minLevel = options?.level;
-    for (const transportOpts of transportOptsArr) {
-      transportOpts.level = getMinLevel(options?.level, transportOpts.level); // General level may override transport level
-      minLevel = getMinLevel(minLevel, transportOpts.level); // Compute the minLevel from general and all transports
-    }
-
+  constructor(
+    private readonly options: Partial<ILoggerOptions> = {},
+    private readonly transports?: winston.transport[]
+  ) {
     this.winston = createLogger({
       level: options?.level || defaultLogLevel,
       defaultMeta: {module: options?.module || ""} as DefaultMeta,
       format: getFormat(options),
-      transports: transportOptsArr.map((transportOpts) => fromTransportOpts(transportOpts)),
+      transports,
       exitOnError: false,
     });
-    this._level = minLevel || defaultLogLevel;
-    // Store for child logger
-    this._options = options;
-    this._transportOptsArr = transportOptsArr;
   }
 
   error(message: string, context?: LogData, error?: Error): void {
@@ -71,38 +55,17 @@ export class WinstonLogger implements ILogger {
     this.createLogEntry(LogLevel.silly, message, context, error);
   }
 
-  profile(message: string, option?: {level: string; message: string}): void {
-    this.winston.profile(message, option);
-  }
-
-  stream(): Writable {
-    throw Error("Not implemented");
-  }
-
   child(options: ILoggerOptions): WinstonLogger {
     // Concat module tags
-    if (options.module) options.module = [this._options.module, options.module].filter(Boolean).join(" ");
-    return new WinstonLogger({...this._options, ...options}, this._transportOptsArr);
+    if (options.module) options.module = [this.options.module, options.module].filter(Boolean).join(" ");
+    return new WinstonLogger({...this.options, ...options}, this.transports);
   }
 
   private createLogEntry(level: LogLevel, message: string, context?: LogData, error?: Error): void {
     // don't propagate if silenced or message level is more detailed than logger level
-    if (logLevelNum[level] > logLevelNum[this._level]) {
-      return;
-    }
+    // if (logLevelNum[level] > logLevelNum[this._level]) {
+    //   return;
+    // }
     this.winston[level](message, {context, error});
   }
-}
-
-/** Return the min LogLevel from multiple transports */
-function getMinLevel(...levelsArg: (LogLevel | undefined)[]): LogLevel {
-  const levels = levelsArg.filter((level): level is LogLevel => Boolean(level));
-
-  // Only if there are no levels to compute min from, consider defaultLogLevel
-  if (levels.length === 0) return defaultLogLevel;
-
-  return levels.reduce(
-    // error: 0, warn: 1, info: 2, ...
-    (minLevel, level) => (logLevelNum[level] > logLevelNum[minLevel] ? level : minLevel)
-  );
 }
